feat: separated the PageObjectsListUI slice into a separate state
