+++ src/app/store/store.ts
@@ -8,6 +8,7 @@ import locatorsSlice from '../../features/locators/locators.slice';
 import customLocatorSlice from '../../features/locators/customLocator.slice';
 import onboardingSlice from '../../features/onboarding/store/onboarding.slice';
 import pageObjectSlice from '../../features/pageObjects/pageObject.slice';
+import pageObjectsListUISlice from '../../features/pageObjects/pageObjectsListUI.slice';
 
 import { scriptNotifier } from '../../pageServices/scriptNotifier';
 import { updateMessageHandler } from '../../pageServices/scriptMessageHandler';
@@ -30,6 +31,7 @@ const rootReducer = {
   locators: undoable(locatorsSlice, { undoType: 'LOCATOR_UNDO', jumpType: 'LOCATOR_JUMP' }),
   pageObject: undoable(pageObjectSlice, { undoType: 'PAGEOBJECT_UNDO' }),
   progressBar: progressBarSlice,
+  pageObjectsListUI: pageObjectsListUISlice,
   onboarding: onboardingSlice,
   customLocator: customLocatorSlice,
 };
@@ -62,7 +64,8 @@ export const store = configureStore({
 store.subscribe(() => updateMessageHandler(store.dispatch, store.getState()));
 store.subscribe(() => updateSocketMessageHandler(store.dispatch, store.getState()));
 
-export type RootState = ReturnType<typeof store.getState>;
 type AppDispatch = typeof store.dispatch;
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
+
+export type RootState = ReturnType<typeof store.getState>;
