+++ src/features/pageObjects/selectors/pageObjectsListUI.selectors.ts
@@ -1,10 +1,10 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { RootState } from '../../../app/store/store';
-import { PageObjectState } from '../types/pageObjectSlice.types';
+import { PageObjectsListUIState } from '../pageObjectsListUI.slice';
 
-export const selectPageObjectsState = (state: RootState) => state.pageObject.present;
+export const selectPageObjectsListUIState = (state: RootState) => state.pageObjectsListUI;
 
 export const selectIsPageObjectsListUIEnabled = createSelector(
-  [selectPageObjectsState],
-  (pageObjectsState: PageObjectState) => pageObjectsState.isPageObjectsListUIEnabled,
+  [selectPageObjectsListUIState],
+  (pageObjectsListUI: PageObjectsListUIState) => pageObjectsListUI.isPageObjectsListUIEnabled,
 );
